refactor: split drive source runtime bands
